added CDFI csv with a row per Florida loan
Downloaded CSV from https://www.cdfifund.gov/sites/cdfi/files/2023-07/FY2021_Data_Documentation_Instruction.zip and restricted it to Florida only. This is the data that was listed as the source of loan_amount in the eodatascape data dictionary https://eodatascape.datakind.org/data-dictionary.  Am interested in why so many investments outside of investment_areas and why so much money goes to 1% of census tracts